requirements.txt
fastapi
uvicorn
pandas
opencv-python
Jinja2==3.0.3
fastapi-sqlalchemy
python-dotenv
flask